+++ src/test/java/pages/SongsPage.java
@@ -6,7 +6,6 @@ import org.openqa.selenium.support.FindBy;
 
 public class SongsPage extends BasePage{
 
-    //private String FirstSongLocator = "[draggable='true'] [class='title']";
     @FindBy(css="[draggable='true'] [class='title']")
     private WebElement FirstSongLocator;
